feat: breaking to multiple components
